+++ integration/helpers/create-fixture.ts
@@ -25,6 +25,65 @@ const __dirname = url.fileURLToPath(new URL(".", import.meta.url));
 const root = path.join(__dirname, "../..");
 const TMP_DIR = path.join(root, ".tmp", "integration");
 
+export async function spawnTestServer({
+  command,
+  regex,
+  validate,
+  env = {},
+  cwd,
+  timeout = 20000,
+}: {
+  command: string[];
+  regex: RegExp;
+  validate?: (matches: RegExpMatchArray) => void | Promise<void>;
+  env?: Record<string, string>;
+  cwd?: string;
+  timeout?: number;
+}): Promise<{ stop: VoidFunction }> {
+  return new Promise((accept, reject) => {
+    let serverProcess = spawn(command[0], command.slice(1), {
+      env: { ...process.env, ...env },
+      cwd,
+      stdio: "pipe",
+    });
+
+    let started = false;
+    let stdout = "";
+    let rejectTimeout = setTimeout(() => {
+      reject(new Error(`Timed out waiting for server to start (${timeout}ms)`));
+    }, timeout);
+
+    serverProcess.stderr.pipe(process.stderr);
+    serverProcess.stdout.on("data", (chunk: Buffer) => {
+      if (started) return;
+      let newChunk = chunk.toString();
+      stdout += newChunk;
+      let match = stdout.match(regex);
+      if (match) {
+        clearTimeout(rejectTimeout);
+        started = true;
+
+        Promise.resolve(validate?.(match))
+          .then(() => {
+            accept({
+              stop: () => {
+                serverProcess.kill();
+              },
+            });
+          })
+          .catch((error: unknown) => {
+            reject(error);
+          });
+      }
+    });
+
+    serverProcess.on("error", (error: unknown) => {
+      clearTimeout(rejectTimeout);
+      reject(error);
+    });
+  });
+}
+
 export interface FixtureInit {
   buildStdio?: Writable;
   files?: { [filename: string]: string };
@@ -45,7 +104,10 @@ export function json(value: JsonObject) {
   return JSON.stringify(value, null, 2);
 }
 
+const defaultTemplateName = "vite-5-template" satisfies TemplateName;
+
 export async function createFixture(init: FixtureInit, mode?: ServerMode) {
+  let templateName = init.templateName ?? defaultTemplateName;
   let projectDir = await createFixtureProject(init, mode);
   let buildPath = url.pathToFileURL(
     path.join(projectDir, "build/server/index.js")
@@ -134,8 +196,22 @@ export async function createFixture(init: FixtureInit, mode?: ServerMode) {
     };
   }
 
-  let app: ServerBuild = await import(buildPath);
-  let handler = createRequestHandler(app, mode || ServerMode.Production);
+  let build: ServerBuild | null = null;
+  type RequestHandler = (request: Request) => Promise<Response>;
+  let handler: RequestHandler;
+  if (templateName.includes("parcel")) {
+    let serverBuild = await import(buildPath);
+    handler = (serverBuild?.requestHandler ??
+      serverBuild?.default?.requestHandler) as RequestHandler;
+    if (!handler) {
+      throw new Error(
+        "Expected a 'requestHandler' export in Parcel server build"
+      );
+    }
+  } else {
+    build = (await import(buildPath)) as ServerBuild;
+    handler = createRequestHandler(build, mode || ServerMode.Production);
+  }
 
   let requestDocument = async (href: string, init?: RequestInit) => {
     let url = new URL(href, "test://test");
@@ -184,8 +260,10 @@ export async function createFixture(init: FixtureInit, mode?: ServerMode) {
   };
 
   return {
+    templateName,
     projectDir,
-    build: app,
+    build,
+    handler,
     isSpaMode: init.spaMode,
     prerender: init.prerender,
     requestDocument,
@@ -210,66 +288,29 @@ export async function createAppFixture(fixture: Fixture, mode?: ServerMode) {
     stop: VoidFunction;
   }> => {
     if (fixture.useReactRouterServe) {
-      return new Promise(async (accept, reject) => {
-        let port = await getPort();
-
-        let nodebin = process.argv[0];
-        let serveProcess = spawn(
-          nodebin,
-          [
-            "node_modules/@react-router/serve/dist/cli.js",
-            "build/server/index.js",
-          ],
-          {
-            env: {
-              ...process.env,
-              NODE_ENV: mode || "production",
-              PORT: port.toFixed(0),
-            },
-            cwd: fixture.projectDir,
-            stdio: "pipe",
+      let port = await getPort();
+      let { stop } = await spawnTestServer({
+        cwd: fixture.projectDir,
+        command: [
+          process.argv[0],
+          "node_modules/@react-router/serve/dist/cli.js",
+          "build/server/index.js",
+        ],
+        env: {
+          NODE_ENV: mode || "production",
+          PORT: port.toFixed(0),
+        },
+        regex: /\[react-router-serve\] http:\/\/localhost:(\d+)\s/,
+        validate: (matches) => {
+          let parsedPort = parseInt(matches[1], 10);
+          if (port !== parsedPort) {
+            throw new Error(
+              `Expected react-router-serve to start on port ${port}, but it started on port ${parsedPort}`
+            );
           }
-        );
-        // Wait for `started at http://localhost:${port}` to be printed
-        // and extract the port from it.
-        let started = false;
-        let stdout = "";
-        let rejectTimeout = setTimeout(() => {
-          reject(
-            new Error("Timed out waiting for react-router-serve to start")
-          );
-        }, 20000);
-        serveProcess.stderr.pipe(process.stderr);
-        serveProcess.stdout.on("data", (chunk) => {
-          if (started) return;
-          let newChunk = chunk.toString();
-          stdout += newChunk;
-          let match: RegExpMatchArray | null = stdout.match(
-            /\[react-router-serve\] http:\/\/localhost:(\d+)\s/
-          );
-          if (match) {
-            clearTimeout(rejectTimeout);
-            started = true;
-            let parsedPort = parseInt(match[1], 10);
-
-            if (port !== parsedPort) {
-              reject(
-                new Error(
-                  `Expected react-router-serve to start on port ${port}, but it started on port ${parsedPort}`
-                )
-              );
-              return;
-            }
-
-            accept({
-              stop: () => {
-                serveProcess.kill();
-              },
-              port,
-            });
-          }
-        });
+        },
       });
+      return { stop, port };
     }
 
     if (fixture.isSpaMode) {
@@ -313,7 +354,30 @@ export async function createAppFixture(fixture: Fixture, mode?: ServerMode) {
       });
     }
 
-    if (!fixture.build) {
+    if (fixture.templateName.includes("parcel")) {
+      let port = await getPort();
+      let { stop } = await spawnTestServer({
+        cwd: fixture.projectDir,
+        command: [process.argv[0], "start.js"],
+        env: {
+          NODE_ENV: mode || "production",
+          PORT: port.toFixed(0),
+        },
+        regex: /Server listening on port (\d+)\s/,
+        validate: (matches) => {
+          let parsedPort = parseInt(matches[1], 10);
+          if (port !== parsedPort) {
+            throw new Error(
+              `Expected Parcel build server to start on port ${port}, but it started on port ${parsedPort}`
+            );
+          }
+        },
+      });
+      return { stop, port };
+    }
+
+    const build = fixture.build;
+    if (!build) {
       return Promise.reject(
         new Error("Cannot start app server without a build")
       );
@@ -327,7 +391,7 @@ export async function createAppFixture(fixture: Fixture, mode?: ServerMode) {
       app.all(
         "*",
         createExpressHandler({
-          build: fixture.build,
+          build,
           mode: mode || ServerMode.Production,
         })
       );
@@ -366,9 +430,9 @@ export async function createFixtureProject(
   init: FixtureInit = {},
   mode?: ServerMode
 ): Promise<string> {
-  let template = init.templateName ?? "vite-5-template";
-  let integrationTemplateDir = path.resolve(__dirname, template);
-  let projectName = `rr-${template}-${Math.random().toString(32).slice(2)}`;
+  let templateName = init.templateName ?? defaultTemplateName;
+  let integrationTemplateDir = path.resolve(__dirname, templateName);
+  let projectName = `rr-${templateName}-${Math.random().toString(32).slice(2)}`;
   let projectDir = path.join(TMP_DIR, projectName);
   let port = init.port ?? (await getPort());
 
@@ -406,12 +470,56 @@ export async function createFixtureProject(
     projectDir
   );
 
-  build(projectDir, init.buildStdio, mode);
+  if (templateName.includes("parcel")) {
+    parcelBuild(projectDir, init.buildStdio, mode);
+  } else {
+    reactRouterBuild(projectDir, init.buildStdio, mode);
+  }
 
   return projectDir;
 }
 
-function build(projectDir: string, buildStdio?: Writable, mode?: ServerMode) {
+function parcelBuild(
+  projectDir: string,
+  buildStdio?: Writable,
+  mode?: ServerMode
+) {
+  let parcelBin = "node_modules/parcel/lib/bin.js";
+
+  let buildArgs: string[] = [parcelBin, "build", "--no-cache"];
+
+  let buildSpawn = spawnSync("node", buildArgs, {
+    cwd: projectDir,
+    env: {
+      ...process.env,
+      NODE_ENV: mode || ServerMode.Production,
+    },
+  });
+
+  // These logs are helpful for debugging. Remove comments if needed.
+  // console.log("spawning node " + buildArgs.join(" ") + ":\n");
+  // console.log("  STDOUT:");
+  // console.log("  " + buildSpawn.stdout.toString("utf-8"));
+  // console.log("  STDERR:");
+  // console.log("  " + buildSpawn.stderr.toString("utf-8"));
+
+  if (buildStdio) {
+    buildStdio.write(buildSpawn.stdout.toString("utf-8"));
+    buildStdio.write(buildSpawn.stderr.toString("utf-8"));
+    buildStdio.end();
+  }
+
+  if (buildSpawn.error || buildSpawn.status) {
+    console.error(buildSpawn.stderr.toString("utf-8"));
+    throw buildSpawn.error || new Error(`Build failed, check the output above`);
+  }
+}
+
+function reactRouterBuild(
+  projectDir: string,
+  buildStdio?: Writable,
+  mode?: ServerMode
+) {
   // We have a "require" instead of a dynamic import in readConfig gated
   // behind mode === ServerMode.Test to make jest happy, but that doesn't
   // work for ESM configs, those MUST be dynamic imports. So we need to